fingerprint: Reset dimlayer HBM + fp press state on error
The way the fingerprint service handles FINGERPRINT_ERROR messages
allows for situations where the user can end up turning off the screen
while notify_fppress is set to 1. This results in situations where the
screen backlight does not turn on at all when turning on the panel,
which is not an ideal user experience.

Fix this by calling `setDimlayerHbm(0)` and `setFpPress(0)` when
handling FINGERPRINT_ERROR messages.

Change-Id: Ic2754249083c30a9b3b150b825357d894814fe84
Signed-off-by: Sandeep P S <sandymankara11@gmail.com>